Now created struct HighsCallback in HighsCallback.h
