+++ src/lp_data/HStruct.h
@@ -43,31 +43,6 @@ struct HighsObjectiveSolution {
   void clear();
 };
 
-struct HighsCallbackDataOut {
-  HighsLogType log_type;
-  HighsInt simplex_iteration_count;
-  HighsObjectiveSolution objective_solution;
-  void clear();
-};
-
-struct HighsCallbackDataIn {
-  bool user_interrupt;
-  void clear();
-};
-
-struct HighsCallback {
-  void (*highs_user_callback)(const int, const char*, void*,
-			      const HighsCallbackDataOut&,
-			      HighsCallbackDataIn&) = nullptr;
-  void* highs_user_callback_data = nullptr;
-  const int num_type = int(NewHighsCallbackType::kMipImprovingSolution) + 1;
-  std::vector<bool> active;
-  HighsCallbackDataOut highs_callback_data_out;
-  HighsCallbackDataIn highs_callback_data_in;
-  bool callbackAction(const NewHighsCallbackType type);
-  void clear();
-};
-
 struct RefactorInfo {
   bool use = false;
   std::vector<HighsInt> pivot_row;
