+++ src/meson.build
@@ -170,6 +170,7 @@ _srcs = [
     'io/HMpsFF.cpp',
     'io/LoadOptions.cpp',
     'lp_data/Highs.cpp',
+    'lp_data/HighsCallback.cpp',
     'lp_data/HighsDebug.cpp',
     'lp_data/HighsInfo.cpp',
     'lp_data/HighsInfoDebug.cpp',
